Fix HTTP Tests in Mono
When testing for a non-existent accept header in the
description document the behaviour on mono and .net is different in
that when setting the header to null in mono will actually result in an
empty string whereas in .net this will result in no accept header being
set. For now the workaround was to extract a new method
`GetJsonWithoutAcceptHeader` and to come back and ensure that we handle
this case correctly in a separate pull request.

Set a more applicable title for persistent subscription feed

Refine the criteria for returning the descritpion document to only No
accept header and requesting the description document